+++ src/expression_executor/specializations/gpu_execute_conjunction.cpp
@@ -16,7 +16,6 @@
 
 // sirius
 #include <expression_executor/gpu_expression_executor.hpp>
-#include <expression_executor/gpu_expression_executor_state.hpp>
 
 // duckdb
 #include <duckdb/common/exception.hpp>
@@ -25,8 +24,9 @@
 
 // cudf
 #include <cudf/binaryop.hpp>
+#include <cudf/cudf_utils.hpp>
 
-namespace sirius::experimental {
+namespace sirius {
 using execute_result = gpu_expression_executor::execute_result;
 
 execute_result gpu_expression_executor::execute(duckdb::BoundConjunctionExpression const& expr,
@@ -49,7 +49,7 @@ execute_result gpu_expression_executor::execute(duckdb::BoundConjunctionExpressi
 
     auto output = execute(*expr.children[0], execution_mode::AST);
 
-    for (idx_t i = 1; i < expr.children.size(); i++) {
+    for (duckdb::idx_t i = 1; i < expr.children.size(); i++) {
       auto child = execute(*expr.children[i], execution_mode::AST);
 
       auto const& output_expr = _ast_tree.emplace<cudf::ast::operation>(
@@ -93,7 +93,7 @@ execute_result gpu_expression_executor::execute(duckdb::BoundConjunctionExpressi
   auto output = execute(*expr.children[0], execution_mode::MATERIALIZE);
   // DuckDB should prune all scalar conjuncts away
   D_ASSERT(!output.is_scalar());
-  for (idx_t i = 1; i < expr.children.size(); i++) {
+  for (duckdb::idx_t i = 1; i < expr.children.size(); i++) {
     auto child = execute(*expr.children[i], execution_mode::MATERIALIZE);
     D_ASSERT(!child.is_scalar());
     auto output_column = cudf::binary_operation(output.get_column_view(),
@@ -107,62 +107,4 @@ execute_result gpu_expression_executor::execute(duckdb::BoundConjunctionExpressi
   return output;
 }
 
-}  // namespace sirius::experimental
-
-namespace duckdb {
-namespace sirius {
-
-std::unique_ptr<GpuExpressionState> GpuExpressionExecutor::InitializeState(
-  const BoundConjunctionExpression& expr, GpuExpressionExecutorState& root)
-{
-  auto result = make_uniq<GpuExpressionState>(expr, root);
-  for (auto& child : expr.children) {
-    result->AddChild(*child);
-  }
-  return std::move(result);
-}
-
-std::unique_ptr<cudf::column> GpuExpressionExecutor::Execute(const BoundConjunctionExpression& expr,
-                                                             GpuExpressionState* state)
-{
-  auto return_type = GetCudfType(expr.return_type);
-
-  // Resolve the children incrementally into the output
-  std::unique_ptr<cudf::column> output_column;
-  for (idx_t i = 0; i < expr.children.size(); i++) {
-    D_ASSERT(state->child_states[i]->expr.return_type == LogicalType::BOOLEAN);
-
-    auto current_result = Execute(*expr.children[i], state->child_states[i].get());
-
-    if (i == 0) {
-      // Nothing to compare
-      output_column = std::move(current_result);
-    } else {
-      // AND/OR current result with output collecte so far
-      switch (expr.GetExpressionType()) {
-        case ExpressionType::CONJUNCTION_AND:
-          output_column = cudf::binary_operation(current_result->view(),
-                                                 output_column->view(),
-                                                 cudf::binary_operator::LOGICAL_AND,
-                                                 return_type,
-                                                 execution_stream,
-                                                 resource_ref);
-          break;
-        case ExpressionType::CONJUNCTION_OR:
-          output_column = cudf::binary_operation(current_result->view(),
-                                                 output_column->view(),
-                                                 cudf::binary_operator::LOGICAL_OR,
-                                                 return_type,
-                                                 execution_stream,
-                                                 resource_ref);
-          break;
-        default: throw InternalException("Execute[Conjunction]: Unknown conjunction type!");
-      }
-    }
-  }
-
-  return std::move(output_column);
-}
-
 }  // namespace sirius
-}  // namespace duckdb
