+++ src/include/legacy/expression_executor/gpu_expression_executor.hpp
@@ -1,0 +1,207 @@
+/*
+ * Copyright 2025, Sirius Contributors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+// sirius
+#include <config.hpp>
+#include <gpu_buffer_manager.hpp>
+#include <gpu_columns.hpp>
+#include <legacy/expression_executor/gpu_expression_executor_state.hpp>
+
+// duckdb
+#include <duckdb/planner/expression.hpp>
+#include <duckdb/planner/expression/bound_between_expression.hpp>
+#include <duckdb/planner/expression/bound_case_expression.hpp>
+#include <duckdb/planner/expression/bound_cast_expression.hpp>
+#include <duckdb/planner/expression/bound_comparison_expression.hpp>
+#include <duckdb/planner/expression/bound_conjunction_expression.hpp>
+#include <duckdb/planner/expression/bound_constant_expression.hpp>
+#include <duckdb/planner/expression/bound_function_expression.hpp>
+#include <duckdb/planner/expression/bound_operator_expression.hpp>
+#include <duckdb/planner/expression/bound_reference_expression.hpp>
+
+// cucascades
+#include <cucascade/data/data_batch.hpp>
+#include <cucascade/data/data_repository_manager.hpp>
+
+// cudf
+#include <cudf/column/column_view.hpp>
+#include <cudf/types.hpp>
+
+// rmm
+#include <rmm/cuda_stream_view.hpp>
+#include <rmm/resource_ref.hpp>
+
+// standard library
+#include <memory>
+#include <vector>
+
+namespace duckdb {
+namespace sirius {
+
+//===----------------------------------------------------------------------===//
+// GpuExpressionExecutor
+//===----------------------------------------------------------------------===//
+
+/**
+ * @brief The GpuExpressionExecutor is responsible for evaluating expressions on the GPU.
+ */
+struct GpuExpressionExecutor {
+  using data_batch              = cucascade::data_batch;
+  using data_repository_manager = cucascade::data_repository_manager<std::shared_ptr<data_batch>>;
+
+  //===----------Constructor/Destructor(s)----------===//
+  /**
+   * @brief Constructs an expression executor with a single expression
+   *
+   * @param expr The expression to evaluate
+   * @param resource_ref The rmm::device_async_resource_ref to pass to cudf APIs for allocations
+   */
+  GpuExpressionExecutor(
+    const Expression& expr,
+    rmm::device_async_resource_ref resource_ref = cudf::get_current_device_resource_ref());
+
+  /**
+   * @brief Constructs an expression executor with a set of expressions
+   *
+   * @param expressions The expressions to evaluate
+   * @param resource_ref The rmm::device_async_resource_ref to pass to cudf APIs for allocations
+   */
+  GpuExpressionExecutor(
+    const vector<unique_ptr<Expression>>& expressions,
+    rmm::device_async_resource_ref resource_ref = cudf::get_current_device_resource_ref());
+
+  //===----------Fields----------===//
+  std::vector<const Expression*> expressions;  ///< The expressions to execute
+  std::vector<std::unique_ptr<GpuExpressionExecutorState>>
+    states;  ///< The execution states associated with each expression to execute
+  std::vector<shared_ptr<GPUColumn>>
+    input_columns;                              ///< The input columns for expression evaluation
+  rmm::device_async_resource_ref resource_ref;  ///< The allocator to pass to cudf APIs
+  cudf::size_type input_count;                  ///< The row count of the input table
+  bool has_null_input_column;                   ///< Whether some input column is null
+  rmm::cuda_stream_view execution_stream;       ///< THe stream in which to execute operations
+
+  //===----------Fields for New Execution Model----------===//
+  bool use_data_batch_apis =
+    false;  ///< Whether to use the data_batch APIs in executing bound references
+  std::vector<std::unique_ptr<cudf::column>>
+    output_columns;              ///< The columns generated by the executed expressions
+  cudf::table_view input_table;  ///< The input table
+
+  //===----------Methods----------===//
+  void AddExpression(const Expression& expr);
+  void ClearExpressions();
+
+  // Set the root state of the executor to the given expression
+  void Initialize(const Expression& expr, GpuExpressionExecutorState& state);
+
+  // Set the input count and columns for the expression executor
+  void SetInputColumns(const GPUIntermediateRelation& input_relation);
+
+  // Before evaluating an expression, check the leaves for nullptrs
+  // (Assumes the input columns have already been set)
+  [[nodiscard]] bool HasNullLeaf(const Expression& expr) const;
+  template <typename ExpressionT>
+  bool HasNullLeafLoop(const ExpressionT& expr) const;
+
+  // Execute the set of expressions with the given input relation and store the result in the
+  // output relation (Provides the main interface with client code for Projections).
+  void Execute(const GPUIntermediateRelation& input_relation,
+               GPUIntermediateRelation& output_relation,
+               rmm::cuda_stream_view stream = rmm::cuda_stream_default);
+
+  // Execute the set of expressions with the given input relation and compact into the output
+  // relation based on the resulting selection vector (Provides the main interface with client
+  // code for Filters).
+  void Select(GPUIntermediateRelation& input_relation,
+              GPUIntermediateRelation& output_relation,
+              rmm::cuda_stream_view stream = rmm::cuda_stream_default);
+
+  /**
+   * @brief Executes the current set of expressions against the given input batch and emits a new
+   * output batch holding the results.
+   *
+   * @param input_batch The input batch against which to evaluate expressions
+   * @param stream The stream in which to execute the operations in the expression tree
+   *
+   * @return std::shared_ptr<cucascade::data_batch> The result of the evaluated expressions
+   *
+   * @note It is required that there is only one boolean expression in the current expression set.
+   */
+  std::shared_ptr<data_batch> execute(std::shared_ptr<data_batch> input_batch,
+                                      rmm::cuda_stream_view stream);
+
+  /**
+   * @brief Evaluates a boolean expression and filters the input batch according to the result.
+   *
+   * @param input_batch The input batch against which to evaluate the expression
+   * @param stream The stream in which to execute the operations in the expression tree
+   *
+   * @return std::shared_ptr<cucascade::data_batch> The input batch filtered by the boolean
+   * expression
+   */
+  std::shared_ptr<cucascade::data_batch> select(std::shared_ptr<data_batch> input_batch,
+                                                rmm::cuda_stream_view stream);
+
+  // Execute the expression at the given index and return the result
+  std::unique_ptr<cudf::column> ExecuteExpression(idx_t expression_idx);
+
+  //----------Execute + Specializations----------//
+  std::unique_ptr<cudf::column> Execute(const Expression& expr, GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundBetweenExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundCaseExpression& expr, GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundCastExpression& expr, GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundComparisonExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundConjunctionExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundConstantExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundFunctionExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundOperatorExpression& expr,
+                                        GpuExpressionState* state);
+  std::unique_ptr<cudf::column> Execute(const BoundReferenceExpression& expr,
+                                        GpuExpressionState* state);
+
+  //===----------Initialize State + Specializations----------===//
+  static std::unique_ptr<GpuExpressionState> InitializeState(const Expression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundBetweenExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundCaseExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundCastExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundComparisonExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundConjunctionExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundConstantExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundFunctionExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundOperatorExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+  static std::unique_ptr<GpuExpressionState> InitializeState(const BoundReferenceExpression& expr,
+                                                             GpuExpressionExecutorState& state);
+};
+
+}  // namespace sirius
+}  // namespace duckdb
