+++ src/legacy/operator/gpu_physical_table_scan.cpp
@@ -26,8 +26,8 @@
 #include "duckdb/planner/expression/bound_reference_expression.hpp"
 #include "duckdb/planner/filter/conjunction_filter.hpp"
 #include "duckdb/planner/filter/constant_filter.hpp"
-#include "expression_executor/gpu_dispatcher.hpp"
-#include "expression_executor/gpu_expression_executor.hpp"
+#include "legacy/expression_executor/gpu_dispatcher.hpp"
+#include "legacy/expression_executor/gpu_expression_executor.hpp"
 #include "gpu_buffer_manager.hpp"
 #include "gpu_columns.hpp"
 #include "log/logging.hpp"
