+++ src/op/scan/sirius_gpu_parquet_scan_operator.cpp
@@ -23,6 +23,7 @@
 
 // cudf
 #include <cudf/io/parquet.hpp>
+#include <cudf/utilities/memory_resource.hpp>
 
 // cucascade
 #include <cucascade/data/data_batch.hpp>
@@ -183,9 +184,10 @@ std::unique_ptr<operator_data> sirius_gpu_parquet_scan_operator::execute(
   if (std::holds_alternative<std::shared_ptr<duckdb::Expression>>(scan_data->filter_expression)) {
     auto& duckdb_expr = std::get<std::shared_ptr<duckdb::Expression>>(scan_data->filter_expression);
     if (duckdb_expr) {
-      duckdb::sirius::GpuExpressionExecutor gpu_expression_executor(*duckdb_expr);
+      sirius::gpu_expression_executor gpu_expression_executor(
+        duckdb_expr.get(), cudf::get_current_device_resource_ref(), stream);
       auto input_batch  = sirius::make_data_batch(std::move(table), _gpu_memory_space);
-      auto output_batch = gpu_expression_executor.select(input_batch, stream);
+      auto output_batch = gpu_expression_executor.select(input_batch);
       if (!output_batch) { return std::make_unique<operator_data>(); }
       table = output_batch->get_data()->cast<cucascade::gpu_table_representation>().release_table();
       SIRIUS_LOG_DEBUG(
