+++ src/op/sirius_physical_filter.cpp
@@ -56,17 +56,8 @@ std::unique_ptr<operator_data> sirius_physical_filter::execute(const operator_da
   auto& input               = dynamic_cast<const pipelineable_operator_data&>(input_data);
   const auto& input_batches = input.get_data_batches();
 
-  sirius::experimental::expression_executor_strategy strategy;
-  if (!sirius::experimental::string_to_strategy(duckdb::Config::EXPRESSION_EXECUTOR_STRATEGY,
-                                                strategy)) {
-    throw duckdb::InvalidInputException(
-      "Invalid expression_executor_strategy '{}'. Valid values: materialize, ast_interpret, "
-      "ast_jit",
-      duckdb::Config::EXPRESSION_EXECUTOR_STRATEGY);
-  }
-  // The executor uses the data_batch API to filter rows according to `expression`.
-  sirius::experimental::gpu_expression_executor gpu_expression_executor(
-    expression.get(), strategy, cudf::get_current_device_resource_ref(), stream);
+  sirius::gpu_expression_executor gpu_expression_executor(
+    expression.get(), cudf::get_current_device_resource_ref(), stream);
 
   std::vector<std::shared_ptr<cucascade::data_batch>> output_batches;
   output_batches.reserve(input_batches.size());
