+++ src/op/sirius_physical_table_scan.cpp
@@ -154,8 +154,9 @@ std::unique_ptr<operator_data> sirius_physical_table_scan::execute(const operato
   }
 
   if (filter_expr != nullptr) {
-    duckdb::sirius::GpuExpressionExecutor gpu_expression_executor(*filter_expr);
-    output_batch = gpu_expression_executor.select(batch_ref, stream);
+    sirius::gpu_expression_executor gpu_expression_executor(
+      filter_expr.get(), cudf::get_current_device_resource_ref(), stream);
+    output_batch = gpu_expression_executor.select(batch_ref);
     if (!output_batch) { return std::make_unique<pipelineable_operator_data>(); }
   } else {
     output_batch = batch_ref;
